pyproject.toml
[project]
name = "super-todo"
version = "0.1.0"
description = "FastAPI-based super todo list API with Supabase backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.2",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "websockets>=12.0",
    "convex>=0.7.0",
    "browser-use>=0.1.0",
    "playwright>=1.40.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "browser-use-sdk>=1.0.2",
    "redis>=5.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
